Make search usable by limiting results to 100.
100 results is enough to display at any time.
There's a performance problem to render mroe than 100.
The searching is fast no matter what, but the rendering is slow.
We could make it bigger, but 100 is good enough and usually there are
less than 100 results unless you're typing the first char or two of what
you want to type.

Test Plan:
- Was experiencing slow laggy typing when searching.
- Now experiencing blazing fast results.